+++ gdtests/confirm_test.gd
@@ -1,0 +1,9 @@
+class_name ConfirmTest
+
+
+func confirm_null_when_null() -> void:
+	Confirm.is_null(null)
+
+
+func confirm_not_null_when_not_null() -> void:
+	Confirm.is_not_null(true)