+++ gdtests/gdtest.gd
@@ -1,4 +1,0 @@
-class_name gdtest
-
-func test_hi():
-    print("hi")